🐛 FIX: Exclude notebooks from modules in virtual environment that are not part of the report
https://jupyterbook.org/structure/configure.html#config-exclude-non-toc-files

By default, Jupyter Book will build all files that are in your book’s folder, regardless of whether they are specified in the Table of Contents.
Note that files that are in hidden folders (e.g. in .github or .venv) will still be built even if they are not specified in the TOC. You should exclude these files explicitly.

related to https://github.com/executablebooks/jupyter-book/issues/1140